Types: Add types to serialize and with-persistence
